# Optimize `v2/groups` request performance