--- a/src/main/scala/mesosphere/marathon/MarathonSchedulerServiceConfig.scala
+++ b/src/main/scala/mesosphere/marathon/MarathonSchedulerServiceConfig.scala
@@ -10,4 +10,11 @@ trait MarathonSchedulerServiceConfig extends ScallopConf {
     hidden = true,
     default = Some(10000))
 
+  lazy val schedulerActionsExecutionContextSize = opt[Int](
+    "scheduler_actions_execution_context_size",
+    default = Some(8),
+    hidden = true,
+    descr = "INTERNAL TUNING PARAMETER: Scheduler actions component's execution context thread pool size"
+  )
+
 }