--- a/src/main/scala/mesosphere/marathon/MarathonSchedulerServiceConfig.scala
+++ b/src/main/scala/mesosphere/marathon/MarathonSchedulerServiceConfig.scala
@@ -12,7 +12,7 @@ trait MarathonSchedulerServiceConfig extends ScallopConf {
 
   lazy val schedulerActionsExecutionContextSize = opt[Int](
     "scheduler_actions_execution_context_size",
-    default = Some(8),
+    default = Some(Runtime.getRuntime().availableProcessors()),
     hidden = true,
     descr = "INTERNAL TUNING PARAMETER: Scheduler actions component's execution context thread pool size"
   )