--- a/src/main/scala/mesosphere/marathon/core/appinfo/AppInfoConfig.scala
+++ b/src/main/scala/mesosphere/marathon/core/appinfo/AppInfoConfig.scala
@@ -7,14 +7,14 @@ trait AppInfoConfig extends ScallopConf {
 
   lazy val appInfoModuleExecutionContextSize = opt[Int](
     "app_info_module_execution_context_size",
-    default = Some(8),
+    default = Some(Runtime.getRuntime().availableProcessors()),
     hidden = true,
     descr = "INTERNAL TUNING PARAMETER: AppInfo module's execution context thread pool size"
   )
 
   lazy val defaultInfoServiceExecutionContextSize = opt[Int](
     "default_info_service_execution_context_size",
-    default = Some(8),
+    default = Some(Runtime.getRuntime().availableProcessors()),
     hidden = true,
     descr = "INTERNAL TUNING PARAMETER: DefaultInfo service's execution context thread pool size"
   )