--- a/src/main/scala/mesosphere/marathon/core/appinfo/impl/AppInfoBaseData.scala
+++ b/src/main/scala/mesosphere/marathon/core/appinfo/impl/AppInfoBaseData.scala
@@ -18,7 +18,7 @@ import org.slf4j.LoggerFactory
 
 import scala.async.Async.{ async, await }
 import scala.collection.immutable.{ Map, Seq }
-import scala.concurrent.Future
+import scala.concurrent.{ ExecutionContext, Future }
 import scala.util.control.NonFatal
 
 // TODO(jdef) pods rename this to something like ResourceInfoBaseData
@@ -28,10 +28,9 @@ class AppInfoBaseData(
     healthCheckManager: HealthCheckManager,
     deploymentService: DeploymentService,
     taskFailureRepository: TaskFailureRepository,
-    groupManager: GroupManager) {
+    groupManager: GroupManager)(implicit ec: ExecutionContext) {
 
   import AppInfoBaseData._
-  import mesosphere.marathon.core.async.ExecutionContexts.global
 
   if (log.isDebugEnabled) log.debug(s"new AppInfoBaseData $this")
 
@@ -71,28 +70,26 @@ class AppInfoBaseData(
     instanceTracker.instancesBySpec()
   }
 
-  def appInfoFuture(app: AppDefinition, embed: Set[AppInfo.Embed]): Future[AppInfo] = {
+  def appInfoFuture(app: AppDefinition, embeds: Set[AppInfo.Embed]): Future[AppInfo] = async {
     val appData = new AppData(app)
-    embed.foldLeft(Future.successful(AppInfo(app))) { (infoFuture, embed) =>
-      infoFuture.flatMap { info =>
-        embed match {
-          case AppInfo.Embed.Counts =>
-            appData.taskCountsFuture.map(counts => info.copy(maybeCounts = Some(counts)))
-          case AppInfo.Embed.Readiness =>
-            readinessChecksByAppFuture.map(checks => info.copy(maybeReadinessCheckResults = Some(checks(app.id))))
-          case AppInfo.Embed.Deployments =>
-            runningDeploymentsByAppFuture.map(deployments => info.copy(maybeDeployments = Some(deployments(app.id))))
-          case AppInfo.Embed.LastTaskFailure =>
-            appData.maybeLastTaskFailureFuture.map { maybeLastTaskFailure =>
-              info.copy(maybeLastTaskFailure = maybeLastTaskFailure)
-            }
-          case AppInfo.Embed.Tasks =>
-            appData.enrichedTasksFuture.map(tasks => info.copy(maybeTasks = Some(tasks)))
-          case AppInfo.Embed.TaskStats =>
-            appData.taskStatsFuture.map(taskStats => info.copy(maybeTaskStats = Some(taskStats)))
-        }
-      }
-    }
+
+    val taskCountsOpt: Option[TaskCounts] = if (embeds.contains(AppInfo.Embed.Counts)) Some(await(appData.taskCountsFuture)) else None
+    val readinessChecksByAppOpt: Option[Map[PathId, Seq[ReadinessCheckResult]]] = if (embeds.contains(AppInfo.Embed.Readiness)) Some(await(readinessChecksByAppFuture)) else None
+    val runningDeploymentsByAppOpt: Option[Map[PathId, Seq[Identifiable]]] = if (embeds.contains(AppInfo.Embed.Deployments)) Some(await(runningDeploymentsByAppFuture)) else None
+    val lastTaskFailureOpt: Option[TaskFailure] = if (embeds.contains(AppInfo.Embed.LastTaskFailure)) await(appData.maybeLastTaskFailureFuture) else None
+    val enrichedTasksOpt: Option[Seq[EnrichedTask]] = if (embeds.contains(AppInfo.Embed.Tasks)) Some(await(appData.enrichedTasksFuture)) else None
+    val taskStatsOpt: Option[TaskStatsByVersion] = if (embeds.contains(AppInfo.Embed.TaskStats)) Some(await(appData.taskStatsFuture)) else None
+
+    val appInfo = AppInfo(
+      app = app,
+      maybeTasks = enrichedTasksOpt,
+      maybeCounts = taskCountsOpt,
+      maybeDeployments = runningDeploymentsByAppOpt.map(_.apply(app.id)),
+      maybeReadinessCheckResults = readinessChecksByAppOpt.map(_.apply(app.id)),
+      maybeLastTaskFailure = lastTaskFailureOpt,
+      maybeTaskStats = taskStatsOpt
+    )
+    appInfo
   }
 
   /**
@@ -104,12 +101,8 @@ class AppInfoBaseData(
   private[this] class AppData(app: AppDefinition) {
     lazy val now: Timestamp = clock.now()
 
-    lazy val instancesByIdFuture: Future[Map[Instance.Id, Instance]] = instancesByRunSpecFuture.map(_.specInstances(app.id)
-      .foldLeft(Map.newBuilder[Instance.Id, Instance]) { (result, instance) => result += instance.instanceId -> instance }
-      .result()
-    )
-
-    lazy val instancesFuture: Future[Seq[Instance]] = instancesByIdFuture.map(_.values.to[Seq])
+    lazy val instancesFuture: Future[Vector[Instance]] = instancesByRunSpecFuture
+      .map(_.specInstances(app.id).toVector)
 
     lazy val healthByInstanceIdFuture: Future[Map[Instance.Id, Seq[Health]]] = {
       log.debug(s"retrieving health counts for app [${app.id}]")