Improved `v2/groups?embed...` requests performance by introducing dedicated thread pools for fetching info instead of using the `ForkJoinPool`. 
Additionally introduces dedicated execution context for `GroupManagerModule`, `StorageModule` and `SchedulerActions` components.

[Benchmark](https://github.com/wg/wrk) results master:
```
Running 1m test @ http://localhost:8080/v2/groups?embed=...
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.17s   354.93ms   2.00s    70.33%
    Req/Sec     0.12      0.33     1.00     87.76%
  237 requests in 1.00m, 217.12MB read
  Socket errors: connect 0, read 0, write 0, timeout 146
Requests/sec:      3.95
Transfer/sec:      3.62MB
``` 

This branch:
```
Running 1m test @ http://localhost:8080/v2/groups?embed=...
  10 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   249.39ms   62.67ms 580.40ms   65.82%
    Req/Sec     4.01      1.82    10.00     85.60%
  2403 requests in 1.00m, 2.15GB read
Requests/sec:     40.00
Transfer/sec:     36.64MB
```
**Summary**: 10x throughput, 5x latency and 5x std. deviation improvement.

JIRA: COPS-2224