@zen-dog 2018-02-07
> Should this thread pool be ideally related to the number of threads available?
> Should we make it configurable? (even if via an undocumented environment variable)?
I don't think there is much value in setting it to `availableProcessors` or whatever.=. We already have chaos creating a thread for each connection, akka dispatcher pools and a bunch of other threads. And when we used to block everywhere our thread count would grow > 1k threads at times. 
But I do agree that it makes sense to make it configurable.

> It seems like we should ideally create the execution context in the module and share it with the pods status controller?
Shared execution context showed 20% bigger response times (for whatever reason).